renamed file and changed bg img
